+++ src/SsgContext.ts
@@ -1,10 +1,16 @@
 import {SsgFile} from "./util/file/SsgFile"
+import {Logger} from "./Logger"
 
 export type BuiltInVars = SsgFile | undefined
 
 export type VarProp<V> = keyof BuiltInVars | keyof V
 
-export interface SsgContext<V = any> {
+export interface SsgContext<V = any> extends Logger {
+  /**
+   * Context identifier. Used in the logs.
+   */
+  name: string
+
   /**
    * The locale to use to format output (dates, messages, etc.).
    */
@@ -19,11 +25,6 @@ export interface SsgContext<V = any> {
    * The file that will be written.
    */
   outputFile: SsgFile
-
-  log: { (message?: any, ...optionalParams: any[]): void; (...data: any[]): void }
-  debug: { (message?: any, ...optionalParams: any[]): void; (...data: any[]): void }
-  warn: { (message?: any, ...optionalParams: any[]): void; (...data: any[]): void }
-  error: { (message?: any, ...optionalParams: any[]): void; (...data: any[]): void }
 
   getVar(varName: VarProp<V>): string | undefined