+++ src/step/CopyStep.test.ts
@@ -5,7 +5,7 @@ describe("CopyStep", () => {
 
   test("copy", async () => {
     const config = {outDir: "out/"}
-    const copies = ["src/test/test.html"]
+    const copies = ["test/test.html"]
     const step = new CopyStep(copies, config)
     const context = new SsgContextImpl("fr", {})
     const result = await step.execute(context)