+++ src/step/content/replace/htaccess/HtAccessToNetlifyRedirectsReplaceCommand.test.ts
@@ -1,5 +1,5 @@
 import {HtAccessToNetlifyRedirectsReplaceCommand} from "./HtAccessToNetlifyRedirectsReplaceCommand"
-import {testUtil} from "../../../../test/TestUtil"
+import {testUtil} from "../../../../../test/TestUtil"
 import {SsgFile} from "../../../../util/file/SsgFile"
 
 describe("HtAccessToNetlifyRedirectsReplaceCommand", () => {