+++ CHANGELOG.md
@@ -13,11 +13,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
     - under `context.` prefix to avoid collision with JS variables.
         - under reserved prefixes for each built-in object (like `context.inputFile`, etc.)
 
+## [1.2.4] - 2022-11-13
+
+### Fixed
+
+- `SsgFile.lang.variants` don't include current `SsgFile.lang.lang`
+
 ## [1.2.3] - 2022-11-13
 
 ### Fixed
 
-- Find `SsgFile.variants` for path with no directory.
+- Find `SsgFile.lang.variants` for path with no directory.
 
 ## [1.2.2] - 2022-11-13
 
@@ -38,21 +44,27 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `SsgFile.lang` changed from a single string to a `{ lang, variants }` so that you can know which variants of a file exist.
 
 ## [1.1.1] - 2022-11-11
+
 ### Fixed
+
 - Parsing of HTML files
 
 ### Added
+
 - ContentStep test.
 
 ## [1.1.0] - 2022-11-11
+
 This release adds API consistency and generalization, as well as a number of tests.
 
 _However it contains a regression that prevents HTML files to be parsed, so use 1.1.1+ instead._
 
 ### Added
+
 - `SsgFile.readOrNew()` to allow creating a SsgFile in memory that doesn't exist on disk.
 
 ### Changed
+
 - `FileInfo` renamed as `SsgFile` to both improve consistency and avoid name collisions.
 - `HtmlFileInfo` renamed as `HtmlSsgFile` for the same reasons.
 - `writeFileInfo()` becomes `ssgFile.write()`
@@ -60,7 +72,9 @@ _However it contains a regression that prevents HTML files to be parsed, so use 
 - `SsgContext.locale` becomes a single string (not an array of strings anymore) as you cannot output for multiple languages at once (use one context per language if you need to do so).
 
 ### Fixed
+
 - Ability to use any variables names in a context.
 
 ### Regressions
+
 - HTML are not parsed anymore.